backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
pydantic==2.9.2
python-dotenv==1.0.1
groq==0.11.0
firebase-admin==6.5.0
httpx==0.27.2
pytesseract==0.3.13
Pillow==10.4.0